+++ juriscraper/opinions/united_states/federal_bankruptcy/bap10.py
@@ -11,7 +11,7 @@ History:
 """
 
 from datetime import date, datetime, timedelta
-from typing import Optional, Tuple
+from typing import Optional
 from urllib.parse import urlencode
 
 from juriscraper.AbstractSite import logger
@@ -66,7 +66,7 @@ class Site(OpinionSiteLinear):
         }
         self.url = f"{self.base_url}?{urlencode(params)}"
 
-    def _download_backwards(self, dates: Tuple[date]) -> None:
+    def _download_backwards(self, dates: tuple[date]) -> None:
         """Make custom date range request
 
         :param dates: (start_date, end_date) tuple