+++ juriscraper/opinions/united_states/state/la.py
@@ -22,7 +22,9 @@ class Site(OpinionSiteLinear):
         self.url = "http://www.lasc.org/CourtActions/%d" % self.year
         self.status = "Published"
 
-    def _download(self, request_dict={}):
+    def _download(self, request_dict=None):
+        if request_dict is None:
+            request_dict = {}
         landing_page = OpinionSiteLinear._download(self, request_dict)
         if self.test_mode_enabled():
             return [self._get_subpage_html_by_page(landing_page)]
@@ -76,10 +78,7 @@ class Site(OpinionSiteLinear):
 
     def _get_judge_above_anchor(self, anchor):
         path = (
-            "./preceding::*["
-            "starts-with(., 'BY ') or "
-            "contains(., 'CURIAM:')"
-            "]"
+            "./preceding::*[starts-with(., 'BY ') or contains(., 'CURIAM:')]"
         )
         try:
             text = anchor.xpath(path)[-1].text_content()